refactor: extract Linkify into shared, dedupe 3 inline copies
Three nearly-identical Linkify implementations had drifted across the
codebase:
  • CheckCard.tsx — used `text-muted` for dimmed links
  • FeedView.tsx — used `text-neutral-500` (raw Tailwind, off-system)
  • CheckDetailSheet.tsx — same as CheckCard but no `dimmed` prop

Consolidated into src/shared/components/Linkify.tsx with `text-muted`
as the dimmed color (matches the design system per CLAUDE.md). All
three call sites import from shared.

EventCard's separate `LinkifyText` is intentionally NOT migrated — it's
URL-only by design (no @mention rendering for event captions, since
events don't have co-authors). Different semantics, leaving in place.

Net: -164 lines.
